### 2019-04-18 #1391 @whedon
[PRE REVIEW]: stingray: A modern Python library for spectral timing
**Submitting author:** @dhuppenkothen (<a href="http://orcid.org/0000-0002-1169-7486">Daniela Huppenkothen</a>)
**Repository:** <a href="https://github.com/StingraySoftware/stingray" target ="_blank">https://github.com/StingraySoftware/stingray</a>
**Version:** v0.1
