### 2019-04-18 #1391 @arfon
:wave: @dhuppenkothen - this PR should fix the paper compilation issues: https://github.com/StingraySoftware/stingray/pull/394
