### 2019-04-18 #1391 @dhuppenkothen
Ah, I was wondering how to stop them from being pinged!

I think Stuart Mumford (`@Cadair`) might be a good reviewer. Also, Nestor Espinoza `@nespinoza` and Matt Pitkin (`@mattpitkin`). Those are the ones from the list of reviewers whose names stand out, but anyone at the intersection of astronomy and Python would be great, I suspect.
