### 2019-05-03 #1393 @Cadair
@dhuppenkothen Really glad to stingray hit a major milestone :smile:

A couple of notes from the review:

1) Should the citation for Astropy be the second paper from 2018, either instead of or in addition to the 2013 paper?
2) A few of your references in the paper seem to be missing links to DOIs.
3) I am not sure all your dependencies are installed automatically by pip so I opened StingraySoftware/stingray#397

I will have a play around with the code and complete my review as soon as I can (my train is about to arrive and I want to push send on this before it does :laughing:)
