### 2019-05-03 #1393 @Cadair
Oh also, do you plan to release the final 0.1 version before this paper is completed?
