### 2019-05-31 #1393 @1313e
@dhuppenkothen For some reason, I was sent the DOI of the AAS Journals submission of the Stingray paper instead of you.
So, here you go: 10.3847/1538-4357/ab258d
