### 2019-06-08 #1393 @arfon
@dhuppenkothen - could you please add the following two fields to the `paper.md` YAML header:

```
aas-doi: 10.3847/1538-4357/ab258d
aas-journal: Astrophysical Journal
```
