### 2019-06-09 #1393 @arfon
@crawfordsm @dhuppenkothen - is the associated AAS paper in ApJ?
