### 2019-06-09 #1393 @Cadair
This all looks great, apart from the queries I have above about dependencies, missing DOIs and the Astropy citation. :smile:
