### 2019-06-10 #1393 @dhuppenkothen
I had a response all typed out and then somehow forgot to actually send it. Thank you @Cadair for the comments!

To respond:
1. We have fixed the citation (we now cite both, as per recommendation from one of the Astropy developers)
2. We have added DOIs where we could. One is a paper in Astronomy and Astrophysics Supplements from 1993, which seems to not have a DOI I can find (the publisher points back to ADS for issues before 1996), and a couple of others are entries in ASCL, which also seem to not have DOIs associated with them.
3. We're iterating with @Cadair on the dependencies, which I think is pretty close to merging.
4. We have released Version 0.1 officially: https://zenodo.org/record/3239519#.XP6iky-ZMWo
5. We have added the reference to the ApJ paper (and yes, the paper is indeed slated to be published in ApJ)

I realized I don't know whether the JOSS paper gets automatically updated from our master branch.  Do I need to complete a separate step to make sure everything gets updated?
