### 2019-06-10 #1393 @arfon
Great! @dhuppenkothen - at this point could you:

- Make a new release of Stingray that includes all of the changes from this review
- Make an archive of the reviewed software in Zenodo/figshare/other service making sure that the Zenodo archive name matches the title of your paper and that the authors are the same as the paper?
- Then update this thread with the DOI of the archive.
