### 2019-06-11 #1393 @dhuppenkothen
Eek, I broke something. As I was preparing the Zenodo archive, I realized that I'd messed up the co-authors compared to the ApJ version. Trying to fix this, I apparently broke the paper? Any advice, @arfon?
