### 2019-06-11 #1393 @arfon
@dhuppenkothen - this PR should fix it https://github.com/StingraySoftware/stingray/pull/412
