### 2019-06-11 #1393 @dhuppenkothen
Thank you, @arfon! I hope this is all correct this time around. I made a new release and synched it to Zenodo.

The Zenodo DOI is [10.5281/zenodo.3242835](https://zenodo.org/record/3242835#.XP73gy-ZMWo).
